avformat/utils: change the duration to int64_t for update_initial_durations

+++ libavformat/utils.c
@@ -1174,7 +1174,7 @@ static void update_initial_timestamps(AVFormatContext *s, int stream_index,
 }
 
 static void update_initial_durations(AVFormatContext *s, AVStream *st,
-                                     int stream_index, int duration)
+                                     int stream_index, int64_t duration)
 {
     AVPacketList *pktl = s->internal->packet_buffer ? s->internal->packet_buffer : s->internal->parse_queue;
     int64_t cur_dts    = RELATIVE_TS_BASE;
